Load initial ConnState; document and fix races with loading/streaming
Specifically, we can double-process events if we don't take into account
the event NID. This happens because we can receive live events before
we have loaded the initial connection state (list of joined rooms). We
base the initial load around an event NID, so we need to make sure to
ignore any live streamed events which are <= the initial load NID.

We could have alternatively loaded the initial connection state and
/then/ register to receive live events, but this means we could drop
events in the gap between loading events and making the register call
which is arguably worse. We could slap a mutex around it all to atomically
do this, but this means that getting pushed new events is tied to
loading (potentially a lot of) state for a single Conn, increasing
lock contention.