Update API to allow multiple lists; extends filters
Filters: added is_dm, is_favourite, is_encrypted, is_invite.
Lists: Request format has been updated to allow multiple
lists to be specified in a single request, complete with
independent sliding windows and filters. Responses now
include a `"list"` integer to indicate which list the
`op` should be applied to. The `"count"` response field
has been changed to a `"counts"` integer array matching
the lists provided in the request. Ranges in a single
list can be updated as the user scrolls, but empty JSON
objects are required in the `"lists"` array to keep the
existing lists unchanged (and not delete them).