Ensure that when timeline limits are expanded we honour them
We cache timelines in-memory per-room. If the initial timeline
request was small, and then a larger timeline limit was used
(say in a room subscription) then we would not fetch more events.
We now invalidate the cache if we don't have enough events AND
there is no create event in the timeline (indicating there should
be more).