Use a new range delta algorithm
Previoulsy the algorithm was very naive and just did literal delta checks.
For example, with an initial window of `[[0,20],[30,40]]` and a new window
of `[[0,20],[25,35]]` the algorithm would calculate that `[25,35]` has been
added and `[30,40]` has been removed. This isn't entirely correct because
`[30,35]` are the same among the two windows. The inability of the algorithm
to detect this meant more rooms than necessary would be invalidated/synced
when slowly scrolling through the room list.

The new sweep line algorithm is O(nlogn) due to sorting and can accurately
detect partial overlaps. This means now only the trailing edge is INVALIDATEd,
and only the leading edge of the window is SYNCed. This is clearly visible in
the client visualisations.