bugfix: don't race on registering user caches with the dispatcher
When there are multiple requests blocked on an initial /sync, we
could race to set up the user cache, resulting in multiple register
calls with the dispatcher, resulting in the error:

  WRN Dispatcher.Register: receiver already registered

This fixes this by using LoadOrStore instead of just Store. Prior
to this, the caches would replace each other resulting in one cache
just never updating.