refactor: re-design how live updates get to connections
Previously there was far too much focus on _room_ updates e.g
events or unread count changes. With the coming of account data,
specifically global account data, we need to have a more extensible
system to provide updates. Remove ConnEvent and replace it with
`caches.Update` and a plethora of sub-interfaces/structs. Whilst
we're at it, consolidate and formalise the behaviour of global/room
caches by passing them in with each update. This actually fixes a
race condition where previously we were grabbing the latest global
room state which may not match that particular update. Prepare to
pass these updates through to extensions.