optimisation: only extract needed events for required_state where possible
Previously, we would only optimise pulling out event types i.e. if you want
state events with types A and B we only pull out all current state with event
type A or B. This falls down when the client wants their own member event,
as m.room.member is the bulk of the current state. This commit optimises the
SQL queries to also take into account the state key asked for, whilst still
supporting wildcards '*' when they are requested.