bugfix: use a map to store joined room tracker info
Previously we used a slice as this is slightly cheaper, but
since Synapse can return multiple join events in the timeline
it could cause a user ID to be present multiple times in a room.

When this happened, it would cause the `UserCache` callbacks to
be invoked twice for every event, clearly not ideal. By using a
set instead, we make sure that we don't add the same user more
than one time.

Ref: https://github.com/matrix-org/synapse/issues/9768