bugfix: update the roomIDToIndex map when rooms are removed
We weren't doing this previously, but things didn't blow up because
we would almost always call resort() shortly afterwards which _would_
update the map with the new sort positions. In some cases this could
cause the lists to be sorting with incorrect index positions, notably:
 - when an invite is retired.
 - when a room no longer meets filter criteria and is removed.

This could be a major source of duplicate rooms.