bugfix: don't deadlock when connections expire
When connections expire, the user cache listener is removed.
This acquires a write lock.

When user cache events are emitted, a read lock is acquired
_and held_ when invoking listeners.

Unfortunately, invoking a listener can cause a connection to
expire in the case of a full buffer, which would then deadlock
and prevent new user cache listeners from being added.