bugfix: fix #56 by sending the correct NID with each live event
Early versions of the proxy tended to send a list of event JSONs
and the "latest" NID for that batch. This then interacted badly
with later code which used these NIDs to determine if the event
in question should be returned to the client or not. We sometimes
filter them out in cases where the "initial" room has already
included this event, e.g room has msgs A,B,C which were pulled in
initially via a DB call, then we receive C down as an update, we
should not include it else we will send back A,B,C,C. By only
sending the "latest" NID, we will filter out other events in
that batch as they are <= the previously seen latest NID.

This was not tested in E2E tests because it relies on slow pollers
which cause >1 timeline event for a single room to arrive. This
may be a cause of flakey tests. We now have an integration test for
this which injected batches of events for the same room and ensures
they are all seen down the connection.

Thanks to @jplatte and @manuroe for helping debug this.