Fix #59: flakey conn_test
The fundamental issue here is that it isn't possible to express
the following in Go:
 - Call this function, and then do this other thing _before the function returns_

We need to use callbacks for that. Unfortunately, this wasn't done
correctly in this test, and we just took a shortcut:
 - start a new goroutine and call function
 - start a 2nd goroutine, sleep a bit, call function

We assumed that the first goroutine had called the function by the time
the sleep had finished. This assumption isn't always true on slow machines,
hence the flake.

The fix is to wait for the first request to get to the conn handler, and
THEN send the 2nd request.