bugfix: always lazy load typing notif members
Previously we would _only_ do this when live streaming, and not
for initially loaded rooms. This just seems to be an oversight
when this feature was landed. We now defer lazy loading members
until after the response has been fully calculated.

We also need to add a distinction between "strict" matching
on `required_state` and not, as the tests only care about the
LL typing member, but we can also see LL _timeline_ members, which
then sometimes flake the test. Since we don't care about the LL
timeline member, just use a looser check.