e2e test: check ignored users' invites are ignored
I expected that this wouldn't be the case today: I thought that if Bob's
poller sees an invite for Alice from a user she's ignored, then the
proxy would tell Alice about the invite anyway.

This seems not to be the case.

I think this is because the proxy only tells Alice about an invite if
OnInvite is called on her UserCache instance. This happens only at
startup and in response to the `V2InviteRoom` pubsub messsage; that
message only gets emitted by Alice's poller(s). But they won't see any
invites from ignored users, because the upstream homeserver has already
filtered them out.

So: there is nothing to fix here. But I have written the test out now,
so we may as well keep it.

Closes #231. Follows #238, #239.