Add SYNCV3_SERVER_PUB environment variable pointing to the public homeserver URL
`SYNCV3_SERVER` continues to be used for connecting to the server.

`SYNCV3_SERVER_PUB` is a new environment variable (defaulting to the
value of `SYNCV3_SERVER`), so that people who are already pointing
`SYNCV3_SERVER` to a public URL do not need to set both.

This change affects the `server` field in the generated `/client/server.json`
file, so that this field points to the public homeserver URL (`SYNCV3_SERVER_PUB`).

Related to https://github.com/spantaleev/matrix-docker-ansible-deploy/issues/2583

Signed-off-by: Slavi Pantaleev <slavi@devture.com>